mcq_contest/requirements.txt
apyori==1.1.2
asgiref==3.8.1
asttokens==2.4.1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
cryptography==44.0.2
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
Django==5.2
django-cors-headers==4.7.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.15.2
dnspython==2.7.0
executing==2.1.0
google-api-core==2.24.2
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
MouseInfo==0.1.3
nest-asyncio==1.6.0
oauthlib==3.2.2
packaging==24.2
parso==0.8.4
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.26.1
protobuf==6.30.2
psutil==6.1.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.11.1
PyMsgBox==1.0.9
pyparsing==3.2.3
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
pytweening==1.2.0
pytz==2025.2
pywhatkit==5.4
pywin32==308
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.16.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.4.0
wcwidth==0.2.13
wikipedia==1.4.0
